Move zhash code to separate module.
