+++ src/algorithms.rs
@@ -2,6 +2,7 @@ pub mod convex_hull;
 pub mod intersection;
 pub mod polygonization;
 pub mod triangulation;
+pub mod zhash;
 
 #[doc(inline)]
 pub use convex_hull::graham_scan::convex_hull;
