+++ src/algorithms/triangulation.rs
@@ -1,1 +1,1 @@
-pub mod zhash;
+pub mod earclip;
